+++ tests/integration/components/power-select/a11y-test.js
@@ -1,11 +1,11 @@
 import { module, test } from 'qunit';
 import { setupRenderingTest } from 'ember-qunit';
-import { render, triggerKeyEvent } from '@ember/test-helpers';
+import { render, triggerEvent, triggerKeyEvent } from '@ember/test-helpers';
 import { hbs } from 'ember-cli-htmlbars';
 import { numbers, groupedNumbers, countriesWithDisabled } from '../constants';
 import { clickTrigger, findContains } from 'ember-power-select/test-support/helpers';
 
-module('Integration | Component | Ember Power Select (Accesibility)', function(hooks) {
+module('Integration | Component | Ember Power Select (Accessibility)', function(hooks) {
   setupRenderingTest(hooks);
 
   test('Single-select: The top-level options list have `role=listbox` and nested lists have `role=group`', async function(assert) {
@@ -40,7 +40,7 @@ module('Integration | Component | Ember Power Select (Accesibility)', function(h
     [].slice.call(nestedOptionList).forEach((e) => assert.dom(e).hasAttribute('role', 'group'));
   });
 
-  test('Single-select: All options have `role=option` except highlighted options which have `role=alert`', async function(assert) {
+  test('Single-select: All options have `role=option`', async function(assert) {
     assert.expect(15);
 
     this.groupedNumbers = groupedNumbers;
@@ -52,16 +52,10 @@ module('Integration | Component | Ember Power Select (Accesibility)', function(h
 
     await clickTrigger();
 
-    [].slice.call(document.querySelectorAll('li.ember-power-select-option[aria-current="true"]')).forEach((e) => {
-      assert.dom(e).hasAttribute('role', 'alert');
-    });
-
-    [].slice.call(document.querySelectorAll('li.ember-power-select-option[aria-current="false"]')).forEach((e) => {
-      assert.dom(e).hasAttribute('role', 'option');
-    });
+    [].slice.call(document.querySelectorAll('.ember-power-select-option')).forEach((e) => assert.dom(e).hasAttribute('role', 'option'));
   });
 
-  test('Multiple-select: All options have `role=option` except highlighted options which have `role=alert`', async function(assert) {
+  test('Multiple-select: All options have `role=option`', async function(assert) {
     assert.expect(15);
 
     this.groupedNumbers = groupedNumbers;
@@ -73,13 +67,7 @@ module('Integration | Component | Ember Power Select (Accesibility)', function(h
 
     await clickTrigger();
 
-    [].slice.call(document.querySelectorAll('li.ember-power-select-option[aria-current="true"]')).forEach((e) => {
-      assert.dom(e).hasAttribute('role', 'alert');
-    });
-
-    [].slice.call(document.querySelectorAll('li.ember-power-select-option[aria-current="false"]')).forEach((e) => {
-      assert.dom(e).hasAttribute('role', 'option');
-    });
+    [].slice.call(document.querySelectorAll('.ember-power-select-option')).forEach((e) => assert.dom(e).hasAttribute('role', 'option'));
   });
 
   test('Single-select: The selected option has `aria-selected=true` and the rest `aria-selected=false`', async function(assert) {
@@ -151,42 +139,6 @@ module('Integration | Component | Ember Power Select (Accesibility)', function(h
     assert.dom(findContains('.ember-power-select-option', 'two')).hasAttribute('aria-current', 'true', 'the second option has now aria-current=false');
   });
 
-  test('Single-select: The highlighted option has `role=alert` and the rest have `role=option`', async function(assert) {
-    assert.expect(4);
-
-    this.numbers = numbers;
-    await render(hbs`
-      <PowerSelect @options={{this.numbers}} @selected={{this.selected}} @onChange={{action (mut this.selected)}} @searchEnabled={{true}} as |option|>
-        {{option}}
-      </PowerSelect>
-    `);
-
-    await clickTrigger();
-    assert.dom(findContains('.ember-power-select-option', 'one')).hasAttribute('role', 'alert', 'the highlighted option has role=alert');
-    assert.dom('.ember-power-select-option[role="option"]').exists({ count: numbers.length - 1 }, 'All other options have role=option');
-    await triggerKeyEvent('.ember-power-select-search-input', 'keydown', 40);
-    assert.dom(findContains('.ember-power-select-option', 'one')).hasAttribute('role', 'option', 'the first option now has role=option');
-    assert.dom(findContains('.ember-power-select-option', 'two')).hasAttribute('role', 'alert', 'the second option now has role=alert');
-  });
-
-  test('Multiple-select: The highlighted option has `role=alert` and the rest `role=option`', async function(assert) {
-    assert.expect(4);
-
-    this.numbers = numbers;
-    await render(hbs`
-      <PowerSelect @options={{this.numbers}} @selected={{this.selected}} @onChange={{action (mut this.selected)}} @searchEnabled={{true}} as |option|>
-        {{option}}
-      </PowerSelect>
-    `);
-
-    await clickTrigger();
-    assert.dom(findContains('.ember-power-select-option', 'one')).hasAttribute('role', 'alert', 'the highlighted option has role=alert');
-    assert.dom('.ember-power-select-option[role="option"]').exists({ count: numbers.length - 1 }, 'All other options have role=option');
-    await triggerKeyEvent('.ember-power-select-search-input', 'keydown', 40);
-    assert.dom(findContains('.ember-power-select-option', 'one')).hasAttribute('role', 'option', 'the first option now has role=option');
-    assert.dom(findContains('.ember-power-select-option', 'two')).hasAttribute('role', 'alert', 'the second option now has role=alert');
-  });
-
   test('Single-select: Options with a disabled field have `aria-disabled=true`', async function(assert) {
     assert.expect(1);
 
@@ -215,8 +167,8 @@ module('Integration | Component | Ember Power Select (Accesibility)', function(h
     assert.dom('.ember-power-select-option[aria-disabled=true]').exists({ count: 3 }, 'Three of them are disabled');
   });
 
-  test('Single-select: The trigger has `role=button` and `aria-owns=<id-of-dropdown>`', async function(assert) {
-    assert.expect(2);
+  test('Single-select: The trigger has `role=button`', async function(assert) {
+    assert.expect(1);
 
     this.numbers = numbers;
     await render(hbs`
@@ -227,11 +179,10 @@ module('Integration | Component | Ember Power Select (Accesibility)', function(h
 
     await clickTrigger();
     assert.dom('.ember-power-select-trigger').hasAttribute('role', 'button', 'The trigger has role button');
-    assert.dom('.ember-power-select-trigger').hasAttribute('aria-owns', /^ember-basic-dropdown-content-ember\d+$/, 'aria-owns points to the dropdown');
   });
 
-  test('Multiple-select: The trigger has `role=button` and `aria-owns=<id-of-dropdown>`', async function(assert) {
-    assert.expect(2);
+  test('Multiple-select: The trigger has `role=button`', async function(assert) {
+    assert.expect(1);
 
     this.numbers = numbers;
     await render(hbs`
@@ -242,7 +193,6 @@ module('Integration | Component | Ember Power Select (Accesibility)', function(h
 
     await clickTrigger();
     assert.dom('.ember-power-select-trigger').hasAttribute('role', 'button', 'The trigger has role button');
-    assert.dom('.ember-power-select-trigger').hasAttribute('aria-owns', /^ember-basic-dropdown-content-ember\d+$/, 'aria-owns points to the dropdown');
   });
 
   test('Single-select: The trigger attribute `aria-expanded` is true when the dropdown is opened', async function(assert) {
@@ -331,34 +281,6 @@ module('Integration | Component | Ember Power Select (Accesibility)', function(h
     await clickTrigger();
     assert.dom('.ember-power-select-trigger-multiple-input').hasAttribute('type', 'search', 'The type of the input is `search`');
     assert.dom('.ember-power-select-trigger-multiple-input').hasAttribute('aria-controls', /^ember-power-select-options-ember\d+$/, 'The `aria-controls` points to the id of the listbox');
-  });
-
-  test('Single-select: The listbox has `aria-controls=<id-of-the-trigger>`', async function(assert) {
-    assert.expect(1);
-
-    this.numbers = numbers;
-    await render(hbs`
-      <PowerSelect @options={{this.numbers}} @selected={{this.selected}} @onChange={{action (mut this.foo)}} as |option|>
-        {{option}}
-      </PowerSelect>
-    `);
-
-    await clickTrigger();
-    assert.dom('.ember-power-select-options').hasAttribute('aria-controls', /^ember-power-select-trigger-ember\d+$/, 'The listbox controls the trigger');
-  });
-
-  test('Multiple-select: The listbox has `aria-controls=<id-of-the-trigger>`', async function(assert) {
-    assert.expect(1);
-
-    this.numbers = numbers;
-    await render(hbs`
-      <PowerSelectMultiple @options={{this.numbers}} @selected={{this.selected}} @onChange={{action (mut this.foo)}} as |option|>
-        {{option}}
-      </PowerSelectMultiple>
-    `);
-
-    await clickTrigger();
-    assert.dom('.ember-power-select-options').hasAttribute('aria-controls', /^ember-power-select-trigger-ember\d+$/, 'The listbox controls the trigger');
   });
 
   test('Multiple-select: The selected elements are <li>s inside an <ul>, and have an item with `role=button` with `aria-label="remove element"`', async function(assert) {
@@ -477,5 +399,146 @@ module('Integration | Component | Ember Power Select (Accesibility)', function(h
     assert.dom('.ember-power-select-trigger').hasAttribute('role', role, 'The `role` was added.');
     this.set('role', undefined);
     assert.dom('.ember-power-select-trigger').hasAttribute('role', 'button', 'The `role` was defaults to `button`.');
+  });
+
+  test('Dropdown with search disabled has proper aria attributes to associate trigger with the options', async function(assert) {
+    assert.expect(3);
+    this.numbers = numbers;
+
+    await render(hbs`
+      <PowerSelect
+        @options={{this.numbers}}
+        @selected={{this.selected}}
+        @searchEnabled={{false}}
+        @onChange={{action (mut this.selected)}}
+        as |number|
+      >
+        {{number}}
+      </PowerSelect>
+    `);
+
+    await clickTrigger();
+
+    assert.dom('.ember-power-select-trigger').hasAttribute('aria-haspopup', 'listbox');
+    assert.dom('.ember-power-select-trigger').hasAttribute('aria-owns', '');
+    assert.dom('.ember-power-select-trigger').hasAttribute('aria-controls', document.querySelector('.ember-power-select-options').id);
+  });
+
+  test('Dropdown with search enabled has proper aria attributes to associate search box with the options', async function(assert) {
+    assert.expect(5);
+    this.numbers = numbers;
+
+    await render(hbs`
+      <PowerSelect
+        @options={{this.numbers}}
+        @selected={{this.selected}}
+        @searchEnabled={{true}}
+        @onChange={{action (mut this.selected)}}
+        as |number|
+      >
+        {{number}}
+      </PowerSelect>
+    `);
+
+    await clickTrigger();
+
+    assert.dom('.ember-power-select-trigger').hasNoAttribute('aria-haspopup');
+    assert.dom('.ember-power-select-trigger').hasNoAttribute('aria-controls');
+    assert.dom('.ember-power-select-search-input').hasAttribute('role', 'combobox');
+    assert.dom('.ember-power-select-search-input').hasAttribute('aria-haspopup', 'listbox');
+    assert.dom('.ember-power-select-search-input').hasAttribute('aria-controls', document.querySelector('.ember-power-select-options').id);
+  });
+
+  test('Trigger has aria-activedescendant attribute for the highlighted option', async function(assert) {
+    assert.expect(5);
+    this.numbers = numbers;
+
+    await render(hbs`
+      <PowerSelect @options={{this.numbers}} @selected={{this.selected}} @onChange={{action (mut this.selected)}} as |number|>
+        {{number}}
+      </PowerSelect>
+    `);
+
+    assert.dom('.ember-power-select-trigger').hasNoAttribute('aria-activedescendant', 'aria-activedescendant is not present when the dropdown is closed');
+
+    await clickTrigger();
+
+    // by default, the first option is highlighted and marked as aria-activedescendant
+    assert.dom('.ember-power-select-option').hasAttribute('aria-current', 'true', 'The first element is highlighted');
+    assert.dom('.ember-power-select-trigger').hasAttribute('aria-activedescendant', document.querySelector('.ember-power-select-option:nth-child(1)').id, 'The first element is the aria-activedescendant');
+
+    await triggerEvent('.ember-power-select-option:nth-child(4)', 'mouseover');
+
+    assert.dom('.ember-power-select-option:nth-child(4)').hasAttribute('aria-current', 'true', 'The 4th element is highlighted');
+    assert.dom('.ember-power-select-trigger').hasAttribute('aria-activedescendant', document.querySelector('.ember-power-select-option:nth-child(4)').id, 'The 4th element is the aria-activedescendant');
+  });
+
+  test('PowerSelectMultiple with search disabled has proper aria attributes', async function(assert) {
+    assert.expect(7);
+    this.numbers = numbers;
+
+    await render(hbs`
+      <PowerSelectMultiple
+        @options={{this.numbers}}
+        @selected={{this.selected}}
+        @searchEnabled={{false}}
+        @onChange={{action (mut this.selected)}}
+        as |number|
+      >
+        {{number}}
+      </PowerSelectMultiple>
+    `);
+
+    assert.dom('.ember-power-select-trigger').hasAttribute('aria-controls', /^ember-power-select-options-ember\d+$/, 'The trigger has aria-controls value');
+    assert.dom('.ember-power-select-trigger').hasNoAttribute('aria-activedescendant', 'aria-activedescendant is not present when the dropdown is closed');
+    assert.dom('.ember-power-select-trigger').hasAttribute('aria-haspopup', 'listbox', 'aria-haspopup is present on the trigger');
+
+    await clickTrigger();
+
+    // by default, the first option is highlighted and marked as aria-activedescendant
+    assert.dom('.ember-power-select-option').hasAttribute('aria-current', 'true', 'The first element is highlighted');
+    assert.dom('.ember-power-select-trigger').hasAttribute('aria-activedescendant', document.querySelector('.ember-power-select-option:nth-child(1)').id, 'The first element is the aria-activedescendant');
+
+    await triggerEvent('.ember-power-select-option:nth-child(4)', 'mouseover');
+
+    assert.dom('.ember-power-select-option:nth-child(4)').hasAttribute('aria-current', 'true', 'The 4th element is highlighted');
+    assert.dom('.ember-power-select-trigger').hasAttribute('aria-activedescendant', document.querySelector('.ember-power-select-option:nth-child(4)').id, 'The 4th element is the aria-activedescendant');
+  });
+
+  test('PowerSelectMultiple with search enabled has proper aria attributes', async function(assert) {
+    assert.expect(11);
+    this.numbers = numbers;
+
+    await render(hbs`
+      <PowerSelectMultiple
+        @options={{this.numbers}}
+        @selected={{this.selected}}
+        @searchEnabled={{true}}
+        @onChange={{action (mut this.selected)}}
+        as |number|
+      >
+        {{number}}
+      </PowerSelectMultiple>
+    `);
+
+    assert.dom('.ember-power-select-trigger').hasNoAttribute('aria-controls', 'The trigger has no aria-controls value');
+    assert.dom('.ember-power-select-trigger').hasNoAttribute('aria-activedescendant', 'aria-activedescendant is not present on the trigger');
+    assert.dom('.ember-power-select-trigger').hasNoAttribute('aria-haspopup', 'aria-haspopup is not present on the trigger');
+
+    await clickTrigger();
+
+    assert.dom('.ember-power-select-trigger-multiple-input').hasAttribute('role', 'combobox', 'Multi select search box has role combobox');
+    assert.dom('.ember-power-select-trigger-multiple-input').hasAttribute('aria-controls', /^ember-power-select-options-ember\d+$/, 'Multi select search box has aria-controls value');
+    assert.dom('.ember-power-select-trigger-multiple-input').hasAttribute('aria-haspopup', 'listbox', 'Multi select search box has aria-haspopup value');
+    assert.dom('.ember-power-select-trigger-multiple-input').hasAttribute('aria-expanded', 'true', 'Multi select search box has aria-expanded value');
+
+    // by default, the first option is highlighted and marked as aria-activedescendant
+    assert.dom('.ember-power-select-option').hasAttribute('aria-current', 'true', 'The first element is highlighted');
+    assert.dom('.ember-power-select-trigger-multiple-input').hasAttribute('aria-activedescendant', document.querySelector('.ember-power-select-option:nth-child(1)').id, 'The first element is the aria-activedescendant');
+
+    await triggerEvent('.ember-power-select-option:nth-child(4)', 'mouseover');
+
+    assert.dom('.ember-power-select-option:nth-child(4)').hasAttribute('aria-current', 'true', 'The 4th element is highlighted');
+    assert.dom('.ember-power-select-trigger-multiple-input').hasAttribute('aria-activedescendant', document.querySelector('.ember-power-select-option:nth-child(4)').id, 'The 4th element is the aria-activedescendant');
   });
 });